backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0